refactor: split utils files and types
