+++ src/utils/document.ts
@@ -1,11 +1,8 @@
-import { getDay, getDate } from 'date-fns';
-import { dateStringToDate, isMatchDateFormat, timeStringToMinute } from './date';
-
 type HTMLElementOrNull = HTMLElement | null;
 
-export const getCellsFromDocument = () => {
+const getListContent = () => {
   const element = document.querySelector('div.list-content');
-  if (!element) return;
+  if (!element) return [];
 
   const rows = Array.from(element.querySelectorAll('div.trow').values());
   const records = rows.map((row) => ({
@@ -26,47 +23,9 @@ export const getCellsFromDocument = () => {
   return records;
 };
 
-export type ContentCells = ReturnType<typeof getCellsFromDocument>;
-
-// NOTE: getCellsFromDocument で同時に行ったら、参照エラーになったので、popup側で呼び出すように修正した
-export const toAttendanceRecords = (params: ContentCells) => {
-  if (!params) return [];
-
-  return params.map(
-    ({
-      date: dateStr,
-      checkIn: checkInStr,
-      checkOut: checkOutStr,
-      breakTime,
-      workingTime,
-      holidayText = '',
-      ...rest
-    }) => {
-      const isHoliday = holidayText.startsWith('全休');
-      const isFirstDay = isMatchDateFormat(dateStr, 'M/d (EEEEE)');
-      // FIXME: month is different
-      const date = dateStringToDate(dateStr, isFirstDay ? 'M/d (EEEEE)' : 'd (EEEEE)');
-      const dayOfWeek = getDay(new Date());
-      const dayOfMonth = getDate(new Date());
-      const checkIn = checkInStr ? dateStringToDate(checkInStr, 'H:mm') : undefined;
-      const checkOut = checkOutStr ? dateStringToDate(checkOutStr, 'H:mm') : undefined;
-      const breakTimeMinute = breakTime ? timeStringToMinute(breakTime) : undefined;
-      const workingTimeMinute = workingTime ? timeStringToMinute(workingTime) : undefined;
-      return {
-        ...rest,
-        isHoliday,
-        holidayText,
-        isFirstDay,
-        date,
-        dayOfWeek,
-        dayOfMonth,
-        checkIn,
-        checkOut,
-        breakTimeMinute,
-        workingTimeMinute,
-      };
-    },
-  );
+export const getAttendanceReportDocument = () => {
+  const listContent = getListContent();
+  return { listContent };
 };
 
-export type AttendanceRecords = ReturnType<typeof toAttendanceRecords>;
+export type AttendanceReportDocument = ReturnType<typeof getAttendanceReportDocument>;
